+++ file-sharing/src/tabs/samba/ui/ShareEditor.vue
@@ -15,8 +15,7 @@ import {
   ValidationResultView,
 } from "@45drives/houston-common-ui";
 import { server } from "@45drives/houston-common-lib";
-import { type SambaShareConfig, newSambaShareConfig } from "@/tabs/samba/data-types";
-import { KeyValueSyntax } from "@45drives/houston-common-lib";
+import { KeyValueSyntax, SambaShareConfig } from "@45drives/houston-common-lib";
 import { BooleanKeyValueSuite } from "@/tabs/samba/ui/BooleanKeyValueSuite"; // TODO: move to common-ui
 import ShareDirectoryInputAndOptions from "@/common/ui/ShareDirectoryInputAndOptions.vue";
 
@@ -45,10 +44,11 @@ const emit = defineEmits<{
 const globalProcessingState = useGlobalProcessingState();
 
 const shareConf = computed<SambaShareConfig>(() =>
-  props.newShare ? newSambaShareConfig() : props.share
+  props.newShare ? SambaShareConfig.makeNew() : props.share
 );
 
 const { tempObject: tempShareConfig, modified, resetChanges } = useTempObjectStaging(shareConf);
+const shareDirectoryOptionsModified = ref(false);
 
 const validationScope = new ValidationScope();
 
@@ -101,17 +101,34 @@ onMounted(async () => {
 const windowsACLsOptions = BooleanKeyValueSuite(() => tempShareConfig.value.advancedOptions, {
   include: {
     "map acl inherit": "yes",
+    "acl_xattr:ignore system acls": "yes",
     "vfs objects": "acl_xattr",
   },
   exclude: {},
 });
 
+const windowsACLsWithLinuxOptions = BooleanKeyValueSuite(
+  () => tempShareConfig.value.advancedOptions,
+  {
+    include: {
+      "map acl inherit": "yes",
+      "vfs objects": "acl_xattr",
+    },
+    exclude: {
+      "acl_xattr:ignore system acls": "yes",
+    },
+  }
+);
+
 const shadowCopyOptions = BooleanKeyValueSuite(() => tempShareConfig.value.advancedOptions, {
   include: {
+    "vfs objects": ["shadow_copy2"],
+  },
+  suggest: {
     "shadow:snapdir": ".zfs/snapshot",
     "shadow:sort": "desc",
     "shadow:format": "%Y-%m-%d-%H%M%S",
-    "vfs objects": "shadow_copy2",
+    "shadow:localtime": "yes",
   },
   exclude: {},
 });
@@ -122,22 +139,28 @@ const macOSSharesOptions = BooleanKeyValueSuite(() => tempShareConfig.value.adva
     "fruit:metadata": "stream",
     "fruit:zero_file_id": "yes",
     "fruit:nfs_aces": "no",
-    "vfs objects": "catia fruit streams_xattr",
+    "vfs objects": ["catia", "fruit", "streams_xattr"],
   },
   exclude: {},
 });
 
 const auditLogsOptions = BooleanKeyValueSuite(() => tempShareConfig.value.advancedOptions, {
   include: {
-    "vfs objects": "full_audit",
+    "vfs objects": ["full_audit"],
+  },
+  suggest: {
     "full_audit:priority": "notice",
     "full_audit:facility": "local5",
-    "full_audit:success": "connect disconnect openat renameat linkat unlinkat",
-    "full_audit:failure": "connect",
+    "full_audit:success": ["connect", "disconnect", "openat", "renameat", "linkat", "unlinkat"],
+    "full_audit:failure": ["connect"],
     "full_audit:prefix": "???%I???%u???%m???%S???%T???",
   },
   exclude: {},
 });
+
+const shareDirectoryInputAndOptionsRef = ref<InstanceType<
+  typeof ShareDirectoryInputAndOptions
+> | null>(null);
 </script>
 
 <template>
@@ -158,8 +181,15 @@ const auditLogsOptions = BooleanKeyValueSuite(() => tempShareConfig.value.advanc
         <InputField v-model="tempShareConfig.description" :placeholder="_('Describe your share')" />
       </InputLabelWrapper>
 
-      <ShareDirectoryInputAndOptions v-model:path="tempShareConfig.path" :disabled="!newShare" allowNonExisting
-        :validationScope />
+      <ShareDirectoryInputAndOptions
+        v-model:path="tempShareConfig.path"
+        :disabled="!newShare"
+        allowNonExisting
+        :validationScope
+        v-model:modified="shareDirectoryOptionsModified"
+        ref="shareDirectoryInputAndOptionsRef"
+        :newShare="newShare ?? false"
+      />
 
       <ToggleSwitchGroup>
         <ToggleSwitch v-model="tempShareConfig.guestOk">
@@ -210,13 +240,21 @@ const auditLogsOptions = BooleanKeyValueSuite(() => tempShareConfig.value.advanc
         <button class="btn btn-secondary" @click="
             () => {
               resetChanges();
+              shareDirectoryInputAndOptionsRef?.resetChanges?.();
               $emit('cancel');
             }
           ">
           {{ _("Cancel") }}
         </button>
-        <button class="btn btn-primary" @click="$emit('apply', tempShareConfig)"
-          :disabled="!validationScope.isValid() || !modified || globalProcessingState !== 0">
+        <button
+          class="btn btn-primary"
+          @click="$emit('apply', tempShareConfig)"
+          :disabled="
+            !validationScope.isValid() ||
+            (!modified && !shareDirectoryOptionsModified) ||
+            globalProcessingState !== 0
+          "
+        >
           {{ _("Apply") }}
         </button>
       </div>